Delete `src/visualization/strategy.rs` and refactor graphing logic
Remove `src/visualization/strategy.rs` and relocate graphing logic into corresponding modules. The visualization implementation for `Strategy`, `Options`, and `Position` is now more modular, using existing traits and utilities for cleaner code.

+++ examples/option_graph.rs
@@ -12,7 +12,7 @@ use std::error::Error;
 fn create_sample_option() -> Options {
     Options::new(
         OptionType::European,
-        Side::Short,
+        Side::Long,
         "AAPL".to_string(),
         100.0,
         ExpirationDate::Days(30.0),
